+++ ms-auth/CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to the
 
 ## Unreleased
 
+
+## 0.4.0.0
+
+Add Bot Framework support
+
+Breaking changes:
+- MSAuth is the only public interface module
+- 'newNoToken' and 'fetchUpdateToken' are not expored anymore from Session and MSAuth, in favor of a single function 'tokenUpdateLoop' which does both the initialization and the refresh loop
+- 'applyDotEnv' and the 'DotEnv' module are gone. Please use the equivalent package 'dotenv-micro'
+
 ## 0.3.0.0
 
 defaultAzureCredential - simplified version of the Microsoft Identity SDK
@@ -15,7 +25,6 @@ defaultAzureCredential - simplified version of the Microsoft Identity SDK
 introduced MSAuth module that re-exports internal functions
 
 Breaking changes:
-
 - module Network.OAuth2.JWT is not exposed anymore
 - OAuthCfg does not contain fields for client ID and secret anymore
 - client ID and client secret can only be loaded from environment variables